Reconstruct all pumpx apis file.
