+++ tee-worker/omni-executor/native-task-handler/src/types.rs
@@ -1,9 +1,9 @@
 use executor_primitives::Hash;
 use parentchain_rpc_client::TransactionStatus;
 use parity_scale_codec::{Decode, Encode};
-use pumpx::pumpx_types::add_wallet::AddWalletResponse;
-use pumpx::pumpx_types::create_transfer_tx::CreateTransferTxResponse;
-use pumpx::pumpx_types::user_connect::UserConnectResponse;
+use pumpx::methods::add_wallet::AddWalletResponse;
+use pumpx::methods::create_transfer_tx::CreateTransferTxResponse;
+use pumpx::methods::user_connect::UserConnectResponse;
 
 #[derive(Encode, Decode, Debug, PartialEq, Eq)]
 pub enum NativeTaskOk {
