+++ tee-worker/omni-executor/pumpx/src/pumpx_api/methods/common.rs
@@ -54,3 +54,15 @@ impl<T: Codec> ApiResponse<T> {
 
 #[derive(Deserialize, Serialize, Encode, Decode, PartialEq, Eq, Debug, Clone, Default)]
 pub struct EmptyResponse {}
+
+// Used in
+// /v3/trade/create_cross_order
+// /v3/trade/cross_fail
+// /v3/trade/create_limit_order/
+#[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct OrderInfoResponseData {
+	pub order_id: Option<u32>,
+}
+
+pub type OrderInfoResponse = ApiResponse<OrderInfoResponseData>;
