+++ tee-worker/omni-executor/pumpx/src/pumpx_api/methods/send_transfer_tx.rs
@@ -1,0 +1,54 @@
+use super::common::ApiResponse;
+use crate::pumpx_api::PumpxApiClient;
+use parity_scale_codec::{Decode, Encode};
+use serde::{Deserialize, Serialize};
+
+// /v3/trade/send_transfer_tx
+#[derive(Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SendTransferTxBody {
+	pub chain_id: u32,
+	pub tx_data: Vec<String>,
+	pub transfer_id: u32,
+}
+
+#[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct SendTransferTxResponseData {
+	pub tx_hash: Option<Vec<String>>,
+}
+
+pub type SendTransferTxResponse = ApiResponse<SendTransferTxResponseData>;
+
+pub async fn send_transfer_tx_impl(
+	client: &PumpxApiClient,
+
+	access_token: &str,
+	body: SendTransferTxBody,
+	language: Option<String>,
+) -> Result<SendTransferTxResponse, Error> {
+	let endpoint = client.base_url.join("v3/trade/send_transfer_tx").unwrap();
+	let response = client
+		.http_client
+		.post(endpoint)
+		.header("X-Language", language.unwrap_or("en".to_string()))
+		.bearer_auth(access_token)
+		.json(&body)
+		.send()
+		.await
+		.map_err(|e| {
+			log::error!("Failed to send send_transfer_tx request: {:?}", e);
+			e
+		})?;
+
+	let status = response.status();
+	let response = response.error_for_status().map_err(|e| {
+		log::error!("send_transfer_tx request failed with status: {}, error: {:?}", status, e);
+		e
+	})?;
+
+	response.json().await.map_err(|e| {
+		log::error!("Failed to parse send_transfer_tx response: {:?}", e);
+		e
+	})
+}
