+++ tee-worker/omni-executor/pumpx/src/pumpx_api/mod.rs
@@ -1,29 +1,43 @@
 use async_trait::async_trait;
 use reqwest::{Client, Error};
 use url::Url;
-pub mod pumpx_methods;
-pub mod pumpx_types;
-pub use pumpx_methods::user_connect::user_connect_impl;
-use pumpx_types::add_wallet::AddWalletResponse;
-use pumpx_types::create_cross_order::CreateCrossOrderBody;
-use pumpx_types::create_limit_order::CreateLimitOrderBody;
-use pumpx_types::create_market_order_tx::{CreateMarketOrderTxBody, CreateMarketOrderTxResponse};
-use pumpx_types::create_market_order_unsigned_tx::{
-	CreateMarketOrderUnsignedTxBody, CreateMarketOrderUnsignedTxResponse,
+pub mod methods;
+use methods::add_wallet::{add_wallet_impl, AddWalletResponse};
+use methods::create_cross_order::{create_cross_order_impl, CreateCrossOrderBody};
+use methods::create_limit_order::{create_limit_order_impl, CreateLimitOrderBody};
+use methods::create_market_order_unsigned_tx::{
+	create_market_order_unsigned_tx_impl, CreateMarketOrderUnsignedTxBody,
+	CreateMarketOrderUnsignedTxResponse,
 };
-use pumpx_types::create_transfer_tx::{CreateTransferTxBody, CreateTransferTxResponse};
-use pumpx_types::create_transfer_unsigned_tx::{
-	CreateTransferUnsignedTxBody, CreateTransferUnsignedTxResponse,
+use methods::cross_fail::{cross_fail_impl, CrossFailBody};
+use methods::get_user_trade_info::{get_user_trade_info_impl, UserTradeInfoResponse};
+use methods::send_order_tx::{send_order_tx_impl, SendOrderTxBody, SendOrderTxResponse};
+use methods::user_connect::{user_connect_impl, UserConnectResponse};
+use methods::verify_google_code::{verify_google_code_impl, GoogleCode, VerifyGoogleCodeResponse};
+
+use methods::create_transfer_unsigned_tx::{
+	create_transfer_unsigned_tx_impl, CreateTransferUnsignedTxBody,
+	CreateTransferUnsignedTxResponse,
 };
-use pumpx_types::cross_fail::CrossFailBody;
-use pumpx_types::get_account_user_id::{GetAccountUserIdParams, GetAccountUserIdResponse};
-use pumpx_types::get_gas_info::{GetGasInfoParams, GetGasInfoResponse};
-use pumpx_types::order_info::OrderInfoResponse;
-use pumpx_types::send_order_tx::{SendOrderTxBody, SendOrderTxResponse};
-use pumpx_types::send_transfer_tx::{SendTransferTxBody, SendTransferTxResponse};
-use pumpx_types::user_connect::{UserConnectBody, UserConnectResponse};
-use pumpx_types::user_trade_info::UserTradeInfoResponse;
-use pumpx_types::verify_google_code::{GoogleCode, VerifyGoogleCodeResponse};
+
+use methods::send_transfer_tx::{
+	send_transfer_tx_impl, SendTransferTxBody, SendTransferTxResponse,
+};
+
+use methods::create_market_order_tx::{
+	create_market_order_tx_impl, CreateMarketOrderTxBody, CreateMarketOrderTxResponse,
+};
+
+use methods::create_transfer_tx::{
+	create_transfer_tx_impl, CreateTransferTxBody, CreateTransferTxResponse,
+};
+
+use methods::get_gas_info::{get_gas_info_impl, GetGasInfoParams, GetGasInfoResponse};
+
+use methods::common::OrderInfoResponse;
+use methods::get_account_user_id::{
+	get_account_user_id_impl, GetAccountUserIdParams, GetAccountUserIdResponse,
+};
 
 const DEFAULT_BASE_URL: &str = "https://api.pumpx.ai";
 
@@ -165,28 +179,7 @@ impl PumpxApi for PumpxApiClient {
 		google_code: String,
 		language: Option<String>,
 	) -> Result<VerifyGoogleCodeResponse, Error> {
-		let endpoint = self.base_url.join("v3/account/verify_google_code").unwrap();
-		let response = self
-			.http_client
-			.post(endpoint)
-			.header("X-Language", language.unwrap_or("en".to_string()))
-			.bearer_auth(access_token)
-			.json(&GoogleCode { google_code })
-			.send()
-			.await
-			.map_err(|e| {
-				log::error!("Failed to send Google code verification request: {:?}", e);
-				e
-			})?;
-		let status = response.status();
-		let response = response.error_for_status().map_err(|e| {
-			log::error!("Google code verification failed with status: {}, error: {:?}", status, e);
-			e
-		})?;
-		response.json().await.map_err(|e| {
-			log::error!("Failed to parse Google code verification response: {:?}", e);
-			e
-		})
+		verify_google_code_impl(self, access_token, google_code, language).await
 	}
 
 	async fn add_wallet(
@@ -194,42 +187,14 @@ impl PumpxApi for PumpxApiClient {
 		access_token: &str,
 		language: Option<String>,
 	) -> Result<AddWalletResponse, Error> {
-		let endpoint = self.base_url.join("v3/account/add_wallet").unwrap();
-		let response = self
-			.http_client
-			.post(endpoint)
-			.header("Content-Length", 0)
-			.header("X-Language", language.unwrap_or("en".to_string()))
-			.bearer_auth(access_token)
-			.send()
-			.await
-			.map_err(|e| {
-				log::error!("Failed to send add_wallet request: {:?}", e);
-				e
-			})?;
-		let status = response.status();
-		let response = response.error_for_status().map_err(|e| {
-			log::error!("add_wallet request failed with status: {}, error: {:?}", status, e);
-			e
-		})?;
-		response.json().await.map_err(|e| {
-			log::error!("Failed to parse add_wallet response: {:?}", e);
-			e
-		})
+		add_wallet_impl(self, access_token, language).await
 	}
 
 	async fn get_user_trade_info(
 		&self,
 		access_token: &str,
 	) -> Result<UserTradeInfoResponse, Error> {
-		let endpoint = self.base_url.join("v3/account/get_user_trade_info").unwrap();
-		self.http_client
-			.get(endpoint)
-			.bearer_auth(access_token)
-			.send()
-			.await?
-			.json()
-			.await
+		get_user_trade_info_impl(self, access_token).await
 	}
 
 	async fn create_market_order_unsigned_tx(
@@ -237,33 +202,7 @@ impl PumpxApi for PumpxApiClient {
 		access_token: &str,
 		body: CreateMarketOrderUnsignedTxBody,
 	) -> Result<CreateMarketOrderUnsignedTxResponse, Error> {
-		let endpoint = self.base_url.join("v3/trade/create_market_order_unsigned_tx").unwrap();
-		let response = self
-			.http_client
-			.post(endpoint)
-			.bearer_auth(access_token)
-			.json(&body)
-			.send()
-			.await
-			.map_err(|e| {
-				log::error!("Failed to send create_market_order_unsigned_tx request: {:?}", e);
-				e
-			})?;
-
-		let status = response.status();
-		let response = response.error_for_status().map_err(|e| {
-			log::error!(
-				"create_market_order_unsigned_tx failed with status: {}, error: {:?}",
-				status,
-				e
-			);
-			e
-		})?;
-
-		response.json().await.map_err(|e| {
-			log::error!("Failed to parse create_market_order_unsigned_tx response: {:?}", e);
-			e
-		})
+		create_market_order_unsigned_tx_impl(self, access_token, body).await
 	}
 
 	async fn send_order_tx(
@@ -271,29 +210,7 @@ impl PumpxApi for PumpxApiClient {
 		access_token: &str,
 		body: SendOrderTxBody,
 	) -> Result<SendOrderTxResponse, Error> {
-		let endpoint = self.base_url.join("v3/trade/send_order_tx").unwrap();
-		let response = self
-			.http_client
-			.post(endpoint)
-			.bearer_auth(access_token)
-			.json(&body)
-			.send()
-			.await
-			.map_err(|e| {
-				log::error!("Failed to send market order transaction: {:?}", e);
-				e
-			})?;
-
-		let status = response.status();
-		let response = response.error_for_status().map_err(|e| {
-			log::error!("Market order transaction failed with status: {}, error: {:?}", status, e);
-			e
-		})?;
-
-		response.json().await.map_err(|e| {
-			log::error!("Failed to parse market order transaction response: {:?}", e);
-			e
-		})
+		send_order_tx_impl(self, access_token, body).await
 	}
 
 	async fn create_limit_order(
@@ -301,15 +218,7 @@ impl PumpxApi for PumpxApiClient {
 		access_token: &str,
 		body: CreateLimitOrderBody,
 	) -> Result<OrderInfoResponse, Error> {
-		let endpoint = self.base_url.join("v3/trade/create_limit_order").unwrap();
-		self.http_client
-			.post(endpoint)
-			.bearer_auth(access_token)
-			.json(&body)
-			.send()
-			.await?
-			.json()
-			.await
+		create_limit_order_impl(self, access_token, body).await
 	}
 
 	async fn create_cross_order(
@@ -317,25 +226,7 @@ impl PumpxApi for PumpxApiClient {
 		access_token: &str,
 		data: CreateCrossOrderBody,
 	) -> Result<OrderInfoResponse, Error> {
-		let endpoint = self.base_url.join("v3/trade/create_cross_order").unwrap();
-		let response = self
-			.http_client
-			.post(endpoint)
-			.bearer_auth(access_token)
-			.json(&data)
-			.send()
-			.await?;
-
-		let status = response.status();
-		let response = response.error_for_status().map_err(|e| {
-			log::error!("Cross order creation failed with status: {}, error: {:?}", status, e);
-			e
-		})?;
-
-		response.json().await.map_err(|e| {
-			log::error!("Failed to parse cross order creation response: {:?}", e);
-			e
-		})
+		create_cross_order_impl(selft, access_token, data).await
 	}
 
 	async fn cross_fail(
@@ -343,25 +234,7 @@ impl PumpxApi for PumpxApiClient {
 		access_token: &str,
 		data: CrossFailBody,
 	) -> Result<OrderInfoResponse, Error> {
-		let endpoint = self.base_url.join("v3/trade/cross_fail").unwrap();
-		let response = self
-			.http_client
-			.post(endpoint)
-			.bearer_auth(access_token)
-			.json(&data)
-			.send()
-			.await?;
-
-		let status = response.status();
-		let response = response.error_for_status().map_err(|e| {
-			log::error!("Cross order failed with status: {}, error: {:?}", status, e);
-			e
-		})?;
-
-		response.json().await.map_err(|e| {
-			log::error!("Failed to parse cross order failed response: {:?}", e);
-			e
-		})
+		cross_fail_impl(self, access_token, data).await
 	}
 
 	#[allow(clippy::too_many_arguments)]
@@ -371,34 +244,7 @@ impl PumpxApi for PumpxApiClient {
 		body: CreateTransferUnsignedTxBody,
 		language: Option<String>,
 	) -> Result<CreateTransferUnsignedTxResponse, Error> {
-		let endpoint = self.base_url.join("v3/trade/create_transfer_unsigned_tx").unwrap();
-		let response = self
-			.http_client
-			.post(endpoint)
-			.header("X-Language", language.unwrap_or("en".to_string()))
-			.bearer_auth(access_token)
-			.json(&body)
-			.send()
-			.await
-			.map_err(|e| {
-				log::error!("Failed to send create_transfer_unsigned_tx request: {:?}", e);
-				e
-			})?;
-
-		let status = response.status();
-		let response = response.error_for_status().map_err(|e| {
-			log::error!(
-				"create_transfer_unsigned_tx request failed with status: {}, error: {:?}",
-				status,
-				e
-			);
-			e
-		})?;
-
-		response.json().await.map_err(|e| {
-			log::error!("Failed to parse create_transfer_unsigned_tx response: {:?}", e);
-			e
-		})
+		create_transfer_unsigned_tx_impl(self, access_token, body, language).await
 	}
 
 	async fn send_transfer_tx(
@@ -407,30 +253,7 @@ impl PumpxApi for PumpxApiClient {
 		body: SendTransferTxBody,
 		language: Option<String>,
 	) -> Result<SendTransferTxResponse, Error> {
-		let endpoint = self.base_url.join("v3/trade/send_transfer_tx").unwrap();
-		let response = self
-			.http_client
-			.post(endpoint)
-			.header("X-Language", language.unwrap_or("en".to_string()))
-			.bearer_auth(access_token)
-			.json(&body)
-			.send()
-			.await
-			.map_err(|e| {
-				log::error!("Failed to send send_transfer_tx request: {:?}", e);
-				e
-			})?;
-
-		let status = response.status();
-		let response = response.error_for_status().map_err(|e| {
-			log::error!("send_transfer_tx request failed with status: {}, error: {:?}", status, e);
-			e
-		})?;
-
-		response.json().await.map_err(|e| {
-			log::error!("Failed to parse send_transfer_tx response: {:?}", e);
-			e
-		})
+		send_transfer_tx_impl(self, access_token, body, language).await
 	}
 
 	async fn create_market_order_tx(
@@ -438,29 +261,7 @@ impl PumpxApi for PumpxApiClient {
 		access_token: &str,
 		body: CreateMarketOrderTxBody,
 	) -> Result<CreateMarketOrderTxResponse, Error> {
-		let endpoint = self.base_url.join("v3/trade/create_market_order_tx").unwrap();
-		let response = self
-			.http_client
-			.post(endpoint)
-			.bearer_auth(access_token)
-			.json(&body)
-			.send()
-			.await
-			.map_err(|e| {
-				log::error!("Failed to send create_market_order_tx request: {:?}", e);
-				e
-			})?;
-
-		let status = response.status();
-		let response = response.error_for_status().map_err(|e| {
-			log::error!("create_market_order_tx failed with status: {}, error: {:?}", status, e);
-			e
-		})?;
-
-		response.json().await.map_err(|e| {
-			log::error!("Failed to parse create_market_order_tx response: {:?}", e);
-			e
-		})
+		create_market_order_tx_impl(self, access_token, body).await
 	}
 
 	#[allow(clippy::too_many_arguments)]
@@ -470,34 +271,7 @@ impl PumpxApi for PumpxApiClient {
 		body: CreateTransferTxBody,
 		language: Option<String>,
 	) -> Result<CreateTransferTxResponse, Error> {
-		let endpoint = self.base_url.join("v3/trade/create_transfer_tx").unwrap();
-		let response = self
-			.http_client
-			.post(endpoint)
-			.header("X-Language", language.unwrap_or("en".to_string()))
-			.bearer_auth(access_token)
-			.json(&body)
-			.send()
-			.await
-			.map_err(|e| {
-				log::error!("Failed to send create_transfer_tx request: {:?}", e);
-				e
-			})?;
-
-		let status = response.status();
-		let response = response.error_for_status().map_err(|e| {
-			log::error!(
-				"create_transfer_tx request failed with status: {}, error: {:?}",
-				status,
-				e
-			);
-			e
-		})?;
-
-		response.json().await.map_err(|e| {
-			log::error!("Failed to parse create_transfer_tx response: {:?}", e);
-			e
-		})
+		create_transfer_tx_impl(self, access_token, body, language).await
 	}
 
 	async fn get_gas_info(
@@ -505,21 +279,10 @@ impl PumpxApi for PumpxApiClient {
 		access_token: &str,
 		chain_id: u32,
 	) -> Result<GetGasInfoResponse, Error> {
-		let endpoint = self.base_url.join("v1/trade/get_gas_info").unwrap();
-		let params = GetGasInfoParams { chain_id };
-		self.http_client
-			.get(endpoint)
-			.bearer_auth(access_token)
-			.query(&params)
-			.send()
-			.await?
-			.json()
-			.await
+		get_gas_info_impl(self, access_token, chain_id).await
 	}
 
 	async fn get_account_user_id(&self, email: String) -> Result<GetAccountUserIdResponse, Error> {
-		let endpoint = self.base_url.join("v3/account/get_account_user_id").unwrap();
-		let params = GetAccountUserIdParams { email };
-		self.http_client.get(endpoint).query(&params).send().await?.json().await
+		get_account_user_id_impl(self, email).await
 	}
 }
