--- tee-worker/omni-executor/pumpx/src/pumpx_api/pumpx_methods/mod.rs
@@ -1,1 +1,0 @@
-pub mod user_connect;
