--- tee-worker/omni-executor/pumpx/src/pumpx_api/pumpx_types/add_wallet.rs
@@ -1,3 +1,0 @@
-use crate::pumpx_types::common::{ApiResponse, EmptyResponse};
-
-pub type AddWalletResponse = ApiResponse<EmptyResponse>;
