--- tee-worker/omni-executor/pumpx/src/pumpx_api/pumpx_types/cross_fail.rs
@@ -1,9 +1,0 @@
-use serde::Serialize;
-
-// /v3/trade/cross_fail
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CrossFailBody {
-	pub request_id: u32,
-	pub fail_reason: String,
-}
