+++ tee-worker/omni-executor/rpc-server/src/methods/omni/request_jwt.rs
@@ -7,7 +7,7 @@ use executor_primitives::OmniAuth;
 use heima_primitives::{Identity, Web2IdentityType};
 use jsonrpsee::RpcModule;
 use native_task_handler::NativeTaskOk;
-use pumpx_types::user_connect::UserConnectResponse;
+use pumpx::methods::user_connect::UserConnectResponse;
 use serde::Serialize;
 
 use super::common::{check_omni_api_response, handle_omni_native_task};
