+++ tee-worker/omni-executor/rpc-server/src/methods/pumpx/add_wallet.rs
@@ -7,7 +7,7 @@ use executor_primitives::OmniAuth;
 use heima_primitives::{Identity, Web2IdentityType};
 use jsonrpsee::RpcModule;
 use native_task_handler::NativeTaskOk;
-use pumpx::pumpx_types::add_wallet::AddWalletResponse;
+use pumpx::methods::add_wallet::AddWalletResponse;
 use serde::Serialize;
 
 use super::common::{check_pumpx_api_response, handle_pumpx_native_task};
