+++ tee-worker/omni-executor/rpc-server/src/methods/pumpx/transfer_widthdraw.rs
@@ -10,7 +10,7 @@ use executor_primitives::OmniAuth;
 use heima_primitives::{Identity, Web2IdentityType};
 use jsonrpsee::RpcModule;
 use native_task_handler::NativeTaskOk;
-use pumpx::pumpx_types::create_transfer_tx::CreateTransferTxResponse;
+use pumpx::methods::create_transfer_tx::CreateTransferTxResponse;
 use serde::Serialize;
 
 use super::common::{check_pumpx_api_response, handle_pumpx_native_task};
